Coinbase Transactions (#12)
* Coinbase Transaction

Code for creating and signing codebase transaction complete. Now need to code for transaction between two addresses.

* Marshal Interface

Defined HexByte type that allows for easier marshaling to json.

* Storing CoinBaseTx to DB

Coinbase transactions can be stored to database by storing their merkle information in the blocks. Next, check if coinbase item exists previously in the chain and handle accordingly.

* Item check if it exists beforehand

New items can only be added to the blockchain if they are not present beforehand. However, since the block needs to contain a merkle tree, blocks without transactions can not be added. Need to discuss if this should be allowed.